+++ src/weapon/weapon.cpp
@@ -415,8 +415,8 @@ const Point2i Weapon::GetGunHolePosition() const
   Point2i pos;
   ActiveCharacter().GetHandPosition(pos);
   Point2i hole(pos + hole_delta * Point2i(ActiveCharacter().GetDirection(),1));
-  double dst = pos.Distance(hole);
-  double angle = pos.ComputeAngle(hole);
+  Double dst = pos.Distance(hole);
+  Double angle = pos.ComputeAngle(hole);
 
   if(ActiveCharacter().GetDirection() == DIRECTION_RIGHT)
     angle += ActiveCharacter().GetFiringAngle();
@@ -477,7 +477,7 @@ void Weapon::UpdateStrength(){
     return ;
 
   uint time = Time::GetInstance()->Read() - m_first_time_loading;
-  double val = (max_strength * time*time) / (MAX_TIME_LOADING*MAX_TIME_LOADING);
+  Double val = (max_strength * time*time) / (MAX_TIME_LOADING*MAX_TIME_LOADING);
 
   m_strength = InRange_Double (val, 0.0, max_strength);
 }
@@ -556,13 +556,13 @@ void Weapon::Draw(){
   {
     if (!EqualsZero(min_angle - max_angle))
     {
-      double angle = m_image->GetRotation_rad();
-      angle += sin( M_PI_2 * double(Time::GetInstance()->Read() - m_time_anim_begin) /(double) ANIM_DISPLAY_TIME) * 2 * M_PI;
+      Double angle = m_image->GetRotation_rad();
+      angle += sin( M_PI_2 * Double(Time::GetInstance()->Read() - m_time_anim_begin) /(Double) ANIM_DISPLAY_TIME) * 2 * M_PI;
       m_image->SetRotation_rad (angle);
     }
     else
     {
-      float scale = sin( 1.5 * M_PI_2 * double(Time::GetInstance()->Read() - m_time_anim_begin) /(double) ANIM_DISPLAY_TIME) / sin(1.5 * M_PI_2);
+      Double scale = sin( 1.5 * M_PI_2 * Double(Time::GetInstance()->Read() - m_time_anim_begin) /(Double) ANIM_DISPLAY_TIME) / sin(1.5 * M_PI_2);
       m_image->Scale(ActiveCharacter().GetDirection() * scale,scale);
 
       if(origin == weapon_origin_OVER) PosXY(x,y); //Recompute position to get the icon centered over the skin
@@ -616,8 +616,8 @@ void Weapon::DrawWeaponFire()
     hole = hole -  Point2i(0, m_weapon_fire->GetHeight()/2);
   else
     hole = hole +  Point2i(0, m_weapon_fire->GetHeight()/2);
-  double dst = hand.Distance(hole);
-  double angle = hand.ComputeAngle(hole);
+  Double dst = hand.Distance(hole);
+  Double angle = hand.ComputeAngle(hole);
 
   angle += ActiveCharacter().GetFiringAngle();
 
@@ -689,8 +689,8 @@ bool Weapon::LoadXml(const xmlNode*  weapon)
   int min_angle_deg = 0, max_angle_deg = 0;
   XmlReader::ReadInt(elem, "min_angle", min_angle_deg);
   XmlReader::ReadInt(elem, "max_angle", max_angle_deg);
-  min_angle = static_cast<double>(min_angle_deg) * M_PI / 180.0;
-  max_angle = static_cast<double>(max_angle_deg) * M_PI / 180.0;
+  min_angle = static_cast<Double>(min_angle_deg) * M_PI / 180.0;
+  max_angle = static_cast<Double>(max_angle_deg) * M_PI / 180.0;
   if(EqualsZero(min_angle - max_angle))
     m_display_crosshair = false;
 
@@ -725,7 +725,7 @@ void Weapon::p_Deselect()
   ActiveCharacter().SetMovement("breathe");
 }
 
-bool Weapon::IsAngleValid(double angle)
+bool Weapon::IsAngleValid(Double angle)
 {
   angle = -angle; // work around incorrect sign
   return min_angle <= angle && angle <= max_angle;
